aadmin login and authentication

+++ research_portal/src/main/java/com/uni/research_portal/service/DepartmentService.java
@@ -3,15 +3,18 @@ package com.uni.research_portal.service;
 import com.uni.research_portal.dto.CreateDepartmentDto;
 import com.uni.research_portal.model.Department;
 import com.uni.research_portal.model.Faculty;
+import com.uni.research_portal.repository.AdminRepository;
 import com.uni.research_portal.repository.DepartmentRepository;
 import com.uni.research_portal.repository.FacultyRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.springframework.web.client.HttpClientErrorException;
 import org.springframework.web.client.RestTemplate;
 
 import java.util.List;
+import static com.uni.research_portal.util.Jwt.extractSubject;
 
 @Service
 public class DepartmentService {
@@ -24,16 +27,29 @@ public class DepartmentService {
     @Autowired
     FacultyRepository facultyRepository;
 
-    public Department createDepartment(CreateDepartmentDto dto){
-        Faculty fac = facultyRepository.findById(dto.getFacultyId()).get();
-        Department newDepartment = new Department(fac, dto.getDepartmentName());
-        departmentRepository.save(newDepartment);
-        return newDepartment;
+    @Autowired
+    AdminRepository adminRepository;
+
+    public Department createDepartment(CreateDepartmentDto dto, String token){
+        if (adminRepository.countByEmail(extractSubject(token))>0){
+            Faculty fac = facultyRepository.findById(dto.getFacultyId()).get();
+            Department newDepartment = new Department(fac, dto.getDepartmentName());
+            departmentRepository.save(newDepartment);
+            return newDepartment;
+        }
+        else{
+            throw new HttpClientErrorException(HttpStatus.UNAUTHORIZED);
+        }
     }
 
-    public ResponseEntity<String> deleteDepartment(int id){
-        departmentRepository.deleteById(id);
-        return new ResponseEntity<>("Department is deleted", HttpStatus.OK);
+    public ResponseEntity<String> deleteDepartment(int id, String token){
+        if (adminRepository.countByEmail(extractSubject(token))>0){
+            departmentRepository.deleteById(id);
+            return new ResponseEntity<>("Department is deleted", HttpStatus.OK);
+        }
+        else{
+            throw new HttpClientErrorException(HttpStatus.UNAUTHORIZED);
+        }
     }
 
     public List<Department> getDepartments(){
